1. feat: add integration-status api 2. refactor: add endpoints for field-mapping feature 3. fix: fix bug with incorrect tab selection on dashboard
