+++ packages/backend/routes/v1/crm/index.ts
@@ -1,8 +1,5 @@
 import express from 'express';
-import { createSession } from 'better-sse';
 import authRouter from './auth';
-import pubsub, { IntegrationStatusSseMessage, PUBSUB_CHANNELS } from '../../../redis/client/pubsub';
-import { logDebug, logError } from '../../../helpers/logger';
 
 const crmRouter = express.Router();
 
@@ -14,22 +11,5 @@ crmRouter.get('/ping', async (_, res) => {
 });
 
 crmRouter.use('/', authRouter);
-
-crmRouter.get('/integration-status/:publicToken', async (req, res) => {
-    try {
-        const publicToken = req.params.publicToken;
-        const { tenantId } = req.query;
-        const session = await createSession(req, res);
-        await pubsub.subscribe(`${PUBSUB_CHANNELS.INTEGRATION_STATUS}_${tenantId}`, async (message: any) => {
-            logDebug('pubsub message', message);
-            let parsedMessage = JSON.parse(message) as IntegrationStatusSseMessage;
-            if (parsedMessage.publicToken === publicToken) {
-                session.push(JSON.stringify(parsedMessage));
-            }
-        });
-    } catch (err: any) {
-        logError(err);
-    }
-});
 
 export default crmRouter;
